+++ test/e2e/spec/util.ts
@@ -356,8 +356,7 @@ const TestUtil = {
       var qr2 = arguments[1];
       var elemID = arguments[2];
       var lfd = LForms.Util.convertFHIRQuestionnaireToLForms(q2, 'R4');
-      var merged = LForms.Util.mergeFHIRDataIntoLForms(
-        'QuestionnaireResponse', qr2, lfd, 'R4');
+      var merged = LForms.Util.mergeFHIRDataIntoLForms(qr2, lfd, 'R4');
 
       // Set a flag so we know when the render is done.
       var formElem = document.getElementById(elemID);