+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
+## [30.0.0-beta.9] 2022-02-24
+### Fixed
+- Fixed a bug that the user saved data was overwritten by the values from initialExpression.
+### Changed
+- Added back the support of a CSS selector as the element parameter for LForms.Util functions 
+  such as getFormData, getUserData and getFormFHIRData.
+- Removed the redundant resourceType parameter in the LForms.Util.mergeFHIRDataIntoLForms function.
+
 ## [30.0.0-beta.8] 2022-02-17
 ### Fixed
 - Fixed a bug that saved user data in a form are lost when answer lists are updated